update procfile, include BigInteger

+++ TWITOFF/models.py
@@ -8,7 +8,7 @@ DB = SQLAlchemy()
 
 class User(DB.Model):
     '''Twitter users that we analyze'''
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     name = DB.Column(DB.String(20), nullable=False)
     newest_tweet_id = DB.Column(DB.BigInteger)
     number_followers = DB.Column(DB.BigInteger)
@@ -20,9 +20,9 @@ def __repr__(self):
 
 class Tweet(DB.Model):
     '''User's tweets from twitter'''
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     text = DB.Column(DB.Unicode(300))
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'), nullable=False)
     user = DB.relationship('User', backref=DB.backref('tweets', lazy=True))
 
     embedding = DB.Column(DB.PickleType, nullable=False)
